+++ bundle/src/main/java/com/adobe/acs/commons/redirects/servlets/ImportRedirectMapServlet.java
@@ -190,9 +190,16 @@ public class ImportRedirectMapServlet extends SlingAllMethodsServlet {
             untilDate.setTime(c3.getDateCellValue());
             props.put(RedirectRule.UNTIL_DATE_PROPERTY_NAME, untilDate);
         }
+        Cell c11 = row.getCell(11);
+        if (DateUtil.isCellDateFormatted(c11)) {
+            Calendar effectiveFrom = Calendar.getInstance();
+            effectiveFrom.setTime(c11.getDateCellValue());
+            props.put(RedirectRule.EFFECTIVE_FROM_PROPERTY_NAME, effectiveFrom);
+        }
         Cell c4 = row.getCell(4);
-        String note = c4 == null ? null : c4.getStringCellValue();
-        props.put(RedirectRule.NOTE_PROPERTY_NAME, note);
+        if(c4 != null) {
+            props.put(RedirectRule.NOTE_PROPERTY_NAME, c4.getStringCellValue());
+        }
         Cell c5 = row.getCell(5);
         boolean ignoreContextPrefix = (c5 != null && c5.getBooleanCellValue());
         props.put(RedirectRule.CONTEXT_PREFIX_IGNORED_PROPERTY_NAME, ignoreContextPrefix);