+++ CHANGELOG.md
@@ -10,6 +10,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 [unreleased changes details]: https://github.com/Adobe-Consulting-Services/acs-aem-commons/compare/acs-aem-commons-5.0.14...HEAD
 
 
+### Added 
+
+- #2977 - Redirect Manager: Add "Effective From" field
+
+### Fixed 
+
+- #2998 - Updated Vanity Path Rewrite Mapper to work on AEM as a Cloud Service
+
 ## 5.4.0 - 2022-10-24
 
 ### Changed